+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.7.1] - 2026-09-26
 ### Added
 - nhl play-by-play: full name for each player ID (e.g. `scoring_player`, `goalie_in_net`) from the play-by-play `rosterSpots`